Improve warning
