+++ domains/rem/src/ui/generatedDates/index.ts
@@ -1,5 +1,4 @@
 export { default as GeneratedDates } from './GeneratedDates';
 export { default as Warning } from './Warning';
-export { default as useLimitsWarning } from './useLimitsWarning';
 
 export * from './types';
