+++ domains/rem/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './date';
 export * from './misc';
 export * from './predicates';
+export * from './useIsCountCapped';
