+++ domains/rem/src/utils/misc.ts
@@ -1,5 +1,7 @@
 import { Frequency } from 'rrule';
 
+import { __ } from '@eventespresso/i18n';
+
 // "freq" is the named capturing group
 const FREQ_PATTERN = /FREQ=(?<freq>(?:YEAR|MONTH|WEEK|DAI)LY)/;
 /**
@@ -52,4 +54,27 @@ export const getMaxDatesLimit = (rruleString: string): number => {
 	const freq = getRecurrenceFrequency(rruleString);
 
 	return DATE_COUNT_LIMITS?.[freq];
+};
+
+export const getLimitsWarning = (rRule: string): string => {
+	const freq = getRecurrenceFrequency(rRule);
+
+	let warning = '';
+
+	switch (true) {
+		case freq === 'YEARLY':
+			warning = __('The number of Event Dates has been capped at 5 for YEARLY recurrence patterns');
+			break;
+		case freq === 'MONTHLY':
+			warning = __('The number of Event Dates has been capped at 36 for MONTHLY recurrence patterns (3 years)');
+			break;
+		case freq === 'WEEKLY':
+			warning = __('The number of Event Dates has been capped at 52 for WEEKLY recurrence patterns (1 year)');
+			break;
+		case freq === 'DAILY':
+			warning = __('The number of Event Dates has been capped at 92 for DAILY recurrence patterns (~3 months)');
+			break;
+	}
+
+	return warning;
 };
